\name{idendr0-package}
\alias{idendr0-package}
\alias{idendr0}
\docType{package}
\title{Interactive dendrograms (backport of 'idendro' to base R graphics and Tcl/Tk GUI)}
\description{Interactive dendrogram that enables the user to select and
color clusters, to zoom and pan the dendrogram, and to visualize
the clustered data not only in a built-in heat map, but also in
GGobi interactive plots and user-supplied plots.}
\details{
\tabular{ll}{Package: \tab idendr0\cr
Type: \tab Package\cr
Title: \tab Interactive dendrograms (backport of 'idendro' to base R graphics and Tcl/Tk GUI)\cr
Version: \tab 1.4.3\cr
Date: \tab 2015-04-10\cr
Author: \tab Tomas Sieger\cr
Depends: \tab tcltk\cr
Suggests: \tab idendro, rggobi, DendSer, cluster, RColorBrewer, testthat\cr
URL: \tab http://github.com/tsieger/idendr0\cr
Maintainer: \tab Tomas Sieger <tomas.sieger@seznam.cz>\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr}
}
\author{Tomas Sieger}

\keyword{ package }


