\name{idendr0-package}
\alias{idendr0-package}
\alias{idendr0}
\docType{package}
\title{Interactive Dendrograms}
\description{Interactive dendrogram that enables the user to select and
color clusters, to zoom and pan the dendrogram, and to visualize
the clustered data not only in a built-in heat map, but also in
GGobi interactive plots and user-supplied plots.
This is a backport of 'idendro' (https://github.com/tsieger/idendro)
to base R graphics and Tcl/Tk GUI.}
\details{
\tabular{ll}{Package: \tab idendr0\cr
Type: \tab Package\cr
Title: \tab Interactive Dendrograms\cr
Version: \tab 1.5.0\cr
Date: \tab 2015-10-06\cr
Author: \tab Tomas Sieger\cr
Depends: \tab tcltk, tkrplot, grDevices\cr
Suggests: \tab rggobi, DendSer, cluster, RColorBrewer, testthat\cr
URL: \tab http://github.com/tsieger/idendr0\cr
Maintainer: \tab Tomas Sieger <tomas.sieger@seznam.cz>\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr}
}
\author{Tomas Sieger}

\keyword{ package }


