\name{idendr0-package}
\alias{idendr0-package}
\alias{idendr0}
\docType{package}
\title{Interactive Dendrograms}
\description{Interactive dendrogram that enables the user to select and
color clusters, to zoom and pan the dendrogram, and to visualize
the clustered data not only in a built-in heat map, but also in
'GGobi' interactive plots and user-supplied plots.
This is a backport of Qt-based 'idendro'
(https://github.com/tsieger/idendro) to base R graphics and Tcl/Tk GUI.}
\details{
\tabular{ll}{Package: \tab idendr0\cr
Type: \tab Package\cr
Title: \tab Interactive Dendrograms\cr
Version: \tab 1.5.2.9000\cr
Date: \tab 2015-10-16\cr
Author: \tab Tomas Sieger\cr
Depends: \tab \cr
Imports: \tab tcltk, tkrplot, grDevices, graphics, stats\cr
Suggests: \tab rggobi, DendSer, cluster, RColorBrewer, datasets\cr
URL: \tab http://github.com/tsieger/idendr0\cr
Maintainer: \tab Tomas Sieger <tomas.sieger@seznam.cz>\cr
BugReports: \tab https://github.com/tsieger/idendr0/issues\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
LazyData: \tab true\cr}
}
\author{Tomas Sieger}

\keyword{ package }


